move code to extpeers
